use abs rect

+++ rosen/modules/render_service/core/pipeline/rs_main_thread.cpp
@@ -3058,7 +3058,7 @@ void RSMainThread::Animate(uint64_t timestamp)
             return false;
         }
         totalAnimationSize += node->animationManager_.GetAnimationsSize();
-        auto frameRateGetFunc = [this](const RSPropertyUnit unit, float velocity, float animationSize) -> int32_t {
+        auto frameRateGetFunc = [this](const RSPropertyUnit unit, float velocity, int32_t animationSize) -> int32_t {
             auto frameRateMgr = HgmCore::Instance().GetFrameRateMgr();
             if (frameRateMgr != nullptr) {
                 return frameRateMgr->GetExpectedFrameRate(unit, velocity, animationSize);
